refactor: add redirect-handler module
